fix: extract a selected folder as one folder, not its loose contents
`ArchiveBatchResolver` expands every selected folder into its descendants
so the 7-Zip engine can extract the subtree by entry index — the zip's
directory entry is a 0-byte marker, so passing only that index yields an
empty folder. `moveExtractedItems` then iterated that same expanded list
and moved *each* entry to `destination/<item.name>`, using the leaf name
and discarding the path.

So extracting MinimalApp.app rained its innards into the target folder and
threw on the first repeated leaf name — `Contents/Resources` and
`Frameworks/Mini.framework/Versions/A/Resources` both want
`destination/Resources`. Since the moves iterate a dictionary, the file
named in the error changed every run, and a parent that happened to move
first left its children's sources gone (NSCocoaErrorDomain 4 instead of
516). Whatever had already moved stayed behind as debris.

Not app-specific: any selected folder was flattened. Drag-out was never
affected because `fulfillDrag` moves the selected item's own temp url in
one `moveItem`, which carries the whole subtree.

Fix: skip entries whose parent is also in the batch. Those are exactly the
descendants expansion added, and they travel with their root's move. The
`itemsById` dictionary the function already builds is the membership test,
so it stays a single pass.

Deriving the roots also beats remembering the pre-expansion selection: if
the user selects both a folder and something inside it, the inner item is
correctly skipped instead of racing its own parent.

Tests move to FolderExtractionTests.swift (EngineTests.swift was at the
file-length limit) and cover the bundle, a plain four-level folder, and a
nested file still landing bare in the destination.

Closes #175
